Step1-Experiment/NetworkFed/requirements.txt
# NetworkFed - Federated Learning Framework Requirements
# =======================================================

# Core ML and Data Science
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# PySyft for Federated Learning
# Note: Install PySyft 0.9.5 manually if needed
syft==0.9.5

# Data Processing
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration and Utilities
pyyaml>=5.4.0
tqdm>=4.62.0
python-dateutil>=2.8.0

# Optional: For advanced features
# jupyter>=1.0.0  # For notebook support
# tensorboard>=2.7.0  # For advanced logging
